--- nav2_depth_estimation_ai/config/perception_pipeline.yaml
@@ -1,45 +1,0 @@
-image_source:
-  type: rgb        # rgb | depth
-  package: usb_cam
-  plugin: usb_cam::UsbCamNode
-  parameters:
-    video_device: /dev/video0
-    image_width: 640
-    image_height: 480
-    pixel_format: mjpeg2rgb
-    frame_rate: 30.0
-  topics:
-    output_topic: /image_raw
-    camera_info_topic: /camera_info
-
-image_preprocessor:
-  enabled: true
-  parameters:
-    resize:
-      width: 640
-      height: 384
-    crop_decimate:
-      x_offset: 0
-      y_offset: 0
-      width: 640
-      height: 480
-      decimation_x: 1
-      decimation_y: 1
-
-depth_estimator:
-  enabled: true
-  package: depth_anything_v3
-  plugin: depth_anything_v3::DepthAnythingV3Node
-  # Model path can be:
-  # - Relative to package share directory
-  # - OR absolute path
-  model_path: models/DA3METRIC-LARGE.onnx
-  model_path_name_argument: 'onnx_path'
-  parameters:
-    precision: fp16
-  topics:
-    input_image: "~/input/image"
-    input_camera_info: "~/input/camera_info"
-    output_depth: "~/output/depth_image"
-
-
